Move friend features to friends page
