+++ frontend/src/components/ProfilePageForm.tsx
@@ -1,10 +1,5 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import User from "../models/User";
-import FriendRequest from "./FriendRequest";
-import dataService from "../services/data";
-import { faUserMinus } from "@fortawesome/free-solid-svg-icons";
-import { faVideo } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Modal from "./Modal";
 
 export interface ProfilePageFormProps {
@@ -17,9 +12,6 @@ export interface ProfilePageFormProps {
 }
 
 function ProfilePageForm(props: ProfilePageFormProps) {
-  const [friends, setFriends] = useState([]);
-  const [friendsRequests, setFriendsRequests] = useState([]);
-  const [refresh, setRefresh] = useState(false);
   const [showDeleteModal, setShowDeleteModal] = useState(false);
 
   const {
@@ -31,53 +23,6 @@ function ProfilePageForm(props: ProfilePageFormProps) {
     deleteUser,
   } = props;
 
-  useEffect(() => {
-    const fetchFriendRequests = async () => {
-      const friendsRequestsResponse = await dataService.fetchData(
-        `/users/${user!.id}/friend-requests`,
-        "GET",
-        {},
-      );
-      setFriendsRequests(friendsRequestsResponse.friends);
-    };
-    fetchFriendRequests();
-  }, [refresh]);
-
-  useEffect(() => {
-    const fetchFriends = async () => {
-      const friendsResponse = await dataService.fetchData(
-        `/users/${user.id}/friends`,
-        "GET",
-        {},
-      );
-      setFriends(friendsResponse.friends);
-    };
-    fetchFriends();
-  }, [refresh]);
-
-  const handleDeclineRequest = async (friend: User) => {
-    await dataService.fetchData(
-      `/users/${user.id}/remove/${friend.id}`,
-      "DELETE",
-      {
-        headers: {
-          "Content-Type": "application/json",
-        },
-      },
-    );
-
-    setRefresh(() => !refresh);
-  };
-
-  const handleAcceptRequest = async (currentId: string) => {
-    await dataService.fetchData(`/users/${user.id}/accept/${currentId}`, "POST", {
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-
-    setRefresh(() => !refresh);
-  };
 
   return (
     <section className="bg-my-darker min-h-screen flex justify-center ">
@@ -182,58 +127,9 @@ function ProfilePageForm(props: ProfilePageFormProps) {
               ></Modal>
             )}
           </div>
-          <h1 className="text-3xl font-bold">Friends:</h1>
-          <hr className="text-my-orange"></hr>
-          <ul>
-            {friends.map((friend: User) => (
-              <li key={friend.id} className="flex flex-row mt-5">
-                <img
-                  src={friend.profile_picture}
-                  className="rounded-full w-20 h-20 border-my-orange border-2 object-cover"
-                />
-                <div className=" ml-5 flex flex-col justify-evenly">
-                  <p className="font-semibold text-2xl">
-                    {friend.first_name} {friend.last_name}
-                  </p>
-                  <div className="flex flex-row">
-                    <button
-                      className={`btn small bg-my-purple text-xs`}
-                      onClick={() => console.log("meeting")}
-                    >
-                      <FontAwesomeIcon icon={faVideo} />
-                    </button>
-                    <button
-                      className={`btn small bg-my-red text-xs`}
-                      onClick={() => handleDeclineRequest(friend)}
-                    >
-                      <FontAwesomeIcon icon={faUserMinus} />
-                    </button>
-                  </div>
-                </div>
-              </li>
-            ))}
-          </ul>
+          
         </div>
-        <div>
-          <h1 className="text-3xl font-bold mt-5">Friend requests:</h1>
-          <hr className="text-my-orange"></hr>
-          <div>
-            {friendsRequests && friendsRequests.length > 0 ? (
-              friendsRequests.map((friend, index) => (
-                <FriendRequest
-                  user={friend}
-                  key={String(index)}
-                  handleAcceptRequest={handleAcceptRequest}
-                  handleDeclineRequest={handleDeclineRequest}
-                />
-              ))
-            ) : (
-              <p className="h1 text-lg">
-                There are currently no friend requests.
-              </p>
-            )}
-          </div>
-        </div>
+        
       </div>
     </section>
   );
